refactor login functions to reduce noise in audit log PR
